requirements.txt
python-dotenv==1.0.1
azure-search-documents==11.4.0
azure-identity==1.17.1
openai==1.84.0
pyodbc==5.2.0